requirements.txt
pymupdf
moviepy
edge-tts
duckduckgo-search
pillow
requests